s/size_t/unsigned/

+++ src/decode.c
@@ -96,7 +96,7 @@ static char8_t const* palm_to_utf8( Byte c ) {
   } // switch
 
   static char8_t utf8_char[ UTF8_CHAR_SIZE_MAX + 1 /*NULL*/ ];
-  size_t len;
+  unsigned len;
 
   if ( cp_is_ascii( cp ) ) {
     if ( !(isspace( (int)cp ) || isprint( (int)cp )) ) {
